Expose CDPClient feature to TreasureData
